Abstract PostListConflictResolver and Helper
- abstract out the conflict resolution logic
- create a PostListHelper for methods shared by that and the PostListEditorPresenter
- add test for PostListHelper date formatter
