+++ WordPress/Classes/ViewRelated/Post/PostListHelper.swift
@@ -1,0 +1,33 @@
+import Foundation
+
+struct PostListHelper {
+    /// Presents the EditPostViewController, with analytics
+    /// - Parameter post: The post to present
+    /// - Parameter loadAutosaveRevision: Creates a new revision
+    /// - Parameter in: An instance of the PostListViewController
+    static func openEditor(with post: Post, loadAutosaveRevision: Bool, in postListViewController: PostListViewController) {
+        let editor = EditPostViewController(post: post, loadAutosaveRevision: loadAutosaveRevision)
+        editor.modalPresentationStyle = .fullScreen
+        postListViewController.present(editor, animated: false)
+        WPAppAnalytics.track(.postListEditAction, withProperties: postListViewController.propertiesForAnalytics(), with: post)
+    }
+
+    /// Returns a date in non-relative time, format 'MMM dd, yyyy @ HH:MM a'
+    static func dateAndTime(for date: Date) -> String {
+        return dateFormatter.string(from: date) + " @ " + timeFormatter.string(from: date)
+    }
+
+    private static let dateFormatter: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.dateStyle = .medium
+        formatter.timeStyle = .none
+        return formatter
+    }()
+
+    private static let timeFormatter: DateFormatter = {
+        let formatter = DateFormatter()
+        formatter.dateStyle = .none
+        formatter.timeStyle = .short
+        return formatter
+    }()
+}
